requirements.txt
langchain>=0.1.0
langchain_community>=0.0.13
langchain_google_genai>=0.0.5
google-generativeai>=0.3.0
faiss-cpu
PyPDF2
numpy
